+++ ride/lifecycle.py
@@ -13,7 +13,7 @@ from ride.metrics import (
     MetricDict,
     MetricMixin,
     OptimisationDirection,
-    make_confusion_matrix,
+    sort_out_figures,
 )
 from ride.utils.gpus import parse_num_gpus
 from ride.utils.logging import getLogger
@@ -104,14 +104,6 @@ class Lifecycle(MetricMixin):
             description="Which gpus should be used. Can be either the number of gpus (e.g. '2') or a list of gpus (e.g. ('0,1').",
         )
         c.add(
-            name="test_confusion_matrix",
-            type=int,
-            default=0,
-            choices=[0, 1],
-            strategy="constant",
-            description="Create and save confusion matrix for test data.",
-        )
-        c.add(
             name="loss",
             type=str,
             default="cross_entropy",
@@ -170,9 +162,14 @@ class Lifecycle(MetricMixin):
         preds, targets = zip(*[(s["pred"], s["target"]) for s in step_outputs])
         preds = torch.vstack([p.unsqueeze(-1) for p in preds]).squeeze(-1)
         targets = torch.vstack([t.unsqueeze(-1) for t in targets]).squeeze(-1)
-        epoch_metrics = prefix_keys(prefix, self.collect_epoch_metrics(preds, targets))
-        metrics = {**mean_step_metrics, **epoch_metrics}
-        LightningModule.log_dict(self, metrics, sync_dist=self._sync_dist)
+        epoch_metrics = prefix_keys(
+            prefix,
+            self.collect_epoch_metrics(preds, targets, prefix.replace("/", "")),
+        )
+        epoch_metrics, epoch_figures = sort_out_figures(epoch_metrics)
+        all_metrics = {**mean_step_metrics, **epoch_metrics}
+        LightningModule.log_dict(self, all_metrics, sync_dist=self._sync_dist)
+        log_figures(self, epoch_figures)
 
     def training_step(self, batch, batch_idx=None):
         if batch_idx == 0:
@@ -194,7 +191,7 @@ class Lifecycle(MetricMixin):
         self.common_epoch_end(step_outputs, prefix="val/")
 
     def test_step(self, batch, batch_idx=None):
-        if not batch:
+        if batch is None:
             return None
 
         x, target = batch[0], batch[1]
@@ -253,37 +250,6 @@ class Lifecycle(MetricMixin):
             ]
 
         self.common_epoch_end(step_outputs, prefix="test/")
-
-        # Make confusion matrix
-        if (
-            self.hparams.test_confusion_matrix
-            and hasattr(self, "num_classes")
-            and self.trainer.progress_bar_callback.is_enabled
-        ):
-            try:
-                preds = torch.tensor(
-                    [p for s in step_outputs for p in s["pred"].argmax(-1)]
-                )
-                targets = torch.tensor([t for s in step_outputs for t in s["target"]])
-
-                fig = make_confusion_matrix(
-                    preds,
-                    targets,
-                    self.num_classes,
-                    count=False,
-                    percent=False,
-                    figsize=(25, 25),
-                    categories=[
-                        f"{c} {i}"
-                        for i, c in enumerate(self.test_dataloader().dataset.classes)
-                    ],
-                    cbar=False,
-                )
-                log_figures(self, {"confusion_matrix": fig})
-            except Exception as e:  # pragma: no cover
-                logger.warning(
-                    f"Unable to save confusion matrix. Caught error: ''{e}''"
-                )
 
 
 def prefix_keys(prefix: str, dictionary: Dict) -> Dict: