+++ ride/runner.py
@@ -12,7 +12,6 @@ from pytorch_lightning.utilities.parsing import AttributeDict
 
 from ride.core import RideDataset, RideModule
 from ride.logging import (
-    CheckpointEveryNSteps,
     ExperimentLoggerCreator,
     ResultsLogger,
     add_experiment_logger,
@@ -80,17 +79,13 @@ class Runner:
                 verbose=True,
                 monitor=f"val/{args.optimization_metric}",  # Comment out when using pl.EvalResult
                 mode=self.Module.metrics()[args.optimization_metric].value,
-                prefix="",
                 save_last=True,
+                every_n_train_steps=args.checkpoint_every_n_steps or None,
             )
         )
         logger.info(
-            f"Checkpointing on val/{args.optimization_metric} with optimisation direction {self.Module.metrics()[args.optimization_metric].value}"
+            f"✅ Checkpointing on val/{args.optimization_metric} with optimisation direction {self.Module.metrics()[args.optimization_metric].value}"
         )
-        if args.checkpoint_every_n_steps:
-            trainer_callbacks.append(
-                CheckpointEveryNSteps(args.checkpoint_every_n_steps)
-            )
         if args.monitor_lr:
             trainer_callbacks.append(LearningRateMonitor(logging_interval="step"))
 
@@ -154,7 +149,7 @@ class Runner:
         if hasattr(self, "trainer"):
             self.trainer.logger = logger
         else:
-            self.trainer = Trainer.from_argparse_args(args, logger=logger)
+            self.trainer = Trainer.from_argparse_args(Namespace(**args), logger=logger)
 
         # Run eval
         self.trainer.test(model, dataloaders[mode]())