+++ CHANGELOG.md
@@ -6,6 +6,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2021-05-17
+### Fixed
+- Pass args correctly to trainer during testing
+
+### Changed
+- CheckpointEveryNSteps now included in ModelCheckpoint c.f. pl==1.3
+- Import from torchmetrics instead of pl.metrics 
+- Moved confusion matrix to RideClassificationDataset and updated plot
+
+### Added
+- Feature extraction and visualisation
+- Lifecycle and Finetuneable mixins always included via RideModule
+- Support for pytorch-lightning==1.3
+- Additional tests: Coverage is now at 92%
+
+### Removed
+- Support for nested inheritance of RideModule
+- Support for pytorch-lightning==1.2
+
+
 ## [0.3.2] - 2021-04-15
 ### Fixed
 - Project dependencies: removed click and added psutil to requirements